PROJECTS/11.Object Detection System/3.ObjectDetection/facebookdetr-resnet-50/requirements.txt
torch
transformers
Pillow
gradio
timm